+++ cinematrix/src/main/java/dev/loslca/cinematrix/model/entity/Role.java
@@ -10,7 +10,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @Column(name = "name", nullable = false, unique = true, length = 50)
     private String name;
@@ -23,8 +23,8 @@ public class Role {
 
     public Role() {}
 
-    public Integer getRoleId() { return roleId; }
-    public void setRoleId(Integer roleId) { this.roleId = roleId; }
+    public Long getRoleId() { return roleId; }
+    public void setRoleId(Long roleId) { this.roleId = roleId; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
